refactor: file output class

+++ src/Outputs/FileOutput.php
@@ -3,23 +3,29 @@
 namespace PHPTools\LaravelDatabaseTask\Outputs;
 
 use PHPTools\LaravelDatabaseTask\Concerns;
-use PHPTools\LaravelDatabaseTask\Contracts\OutputInterface;
+use PHPTools\LaravelDatabaseTask\Contracts;
 
-class FileOutput extends \SplFileObject implements OutputInterface
+class FileOutput extends \SplFileObject implements Contracts\OutputInterface
 {
-    use Concerns\InteractsWithOutput;
+    use Concerns\InteractsWithOutput {
+        getValue as baseGetValue;
+    }
 
     protected bool $autoDelete = true;
 
     public function __destruct()
     {
-        if ($this->autoDelete) {
+        if ($this->autoDelete && $this->isWritable()) {
             @\unlink($this->getRealPath());
         }
     }
 
     public function getValue(): \SplFileObject
     {
+        if (isset($this->value)) {
+            return $this->baseGetValue();
+        }
+
         return $this;
     }
